# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fuse AI System Manager API
  version: 0.1.0
  description: >
    Fuse is an open-source AI system manager exposing three compatibility layers
    (Ollama, OpenAI, Anthropic) plus native system endpoints.  A single Rust
    binary replaces Ollama + vLLM and runs on any device from Raspberry Pi to
    data center.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Fuse Maintainers
servers:
  - url: http://localhost:11434
    description: Local development server (default Ollama-compatible port)
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key passed via the X-API-Key header.
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token passed via the Authorization header.
  headers:
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window.
      schema:
        type: integer
    X-RateLimit-Reset:
      description: UTC epoch seconds when the rate-limit window resets.
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
          description: Machine-readable error code.
          examples: ["rate_limit_exceeded"]
        message:
          type: string
          description: Human-readable error description.
          examples: ["Rate limit exceeded. Please try again later."]
        remediation:
          type: string
          description: Suggested action to resolve the error.
          examples: ["Wait 30 seconds and retry."]

    # ------------------------------------------------------------------
    # Ollama schemas
    # ------------------------------------------------------------------
    OllamaGenerateRequest:
      type: object
      required: [model, prompt]
      properties:
        model:
          type: string
          examples: ["llama3:8b"]
        prompt:
          type: string
        stream:
          type: boolean
          default: false
        system:
          type: string
        template:
          type: string
        context:
          type: array
          items:
            type: integer
            format: uint32
        options:
          type: object
          description: Model-specific options (temperature, top_k, etc.).
    OllamaGenerateResponse:
      type: object
      required: [model, created_at, response, done]
      properties:
        model:
          type: string
        created_at:
          type: string
          format: date-time
        response:
          type: string
        done:
          type: boolean
        context:
          type: array
          items:
            type: integer
            format: uint32
        total_duration:
          type: integer
          format: uint64
        load_duration:
          type: integer
          format: uint64
        prompt_eval_count:
          type: integer
          format: uint32
        eval_count:
          type: integer
          format: uint32
        eval_duration:
          type: integer
          format: uint64
    OllamaChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant]
        content:
          type: string
        images:
          type: array
          items:
            type: string
            description: Base64-encoded image data.
    OllamaChatRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OllamaChatMessage"
        stream:
          type: boolean
          default: false
        options:
          type: object
    OllamaChatResponse:
      type: object
      required: [model, created_at, message, done]
      properties:
        model:
          type: string
        created_at:
          type: string
          format: date-time
        message:
          $ref: "#/components/schemas/OllamaChatMessage"
        done:
          type: boolean
        total_duration:
          type: integer
          format: uint64
        eval_count:
          type: integer
          format: uint32
        eval_duration:
          type: integer
          format: uint64
    ModelDetails:
      type: object
      required: [format, family, parameter_size, quantization_level]
      properties:
        format:
          type: string
          examples: ["gguf"]
        family:
          type: string
        families:
          type: array
          items:
            type: string
        parameter_size:
          type: string
          examples: ["7B"]
        quantization_level:
          type: string
          examples: ["Q4_K_M"]
...
    OllamaPullRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          examples: ["llama3:8b"]
        insecure:
          type: boolean
          default: false
        stream:
          type: boolean
          default: false
    OllamaPullProgress:
      type: object
      required: [status]
      properties:
        status:
          type: string
        digest:
          type: string
        total:
          type: integer
          format: uint64
        completed:
          type: integer
          format: uint64
    ShowResponse:
      type: object
      required: [modelfile, parameters, template, details]
      properties:
        modelfile:
          type: string
        parameters:
          type: string
        template:
          type: string
        details:
          $ref: "#/components/schemas/ModelDetails"
    OllamaEmbeddingsRequest:
      type: object
      required: [model, prompt]
      properties:
        model:
          type: string
        prompt:
          type: string
        options:
          type: object
    OllamaEmbeddingsResponse:
      type: object
      required: [embedding]
      properties:
        embedding:
          type: array
          items:
            type: number
            format: float

    # ------------------------------------------------------------------
    # OpenAI schemas
    # ------------------------------------------------------------------
...
    EmbeddingRequest:
      type: object
      required: [model, input]
      properties:
        model:
          type: string
          examples: ["text-embedding-ada-002"]
        input:
          description: A string or array of strings to embed.
        encoding_format:
          type: string
          enum: [float, base64]
    EmbeddingData:
      type: object
      required: [object, embedding, index]
      properties:
        object:
          type: string
          enum: [embedding]
        embedding:
          type: array
          items:
            type: number
            format: float
        index:
          type: integer
          format: uint32
    EmbeddingUsage:
      type: object
      required: [prompt_tokens, total_tokens]
      properties:
        prompt_tokens:
          type: integer
          format: uint32
        total_tokens:
          type: integer
          format: uint32
    EmbeddingResponse:
      type: object
      required: [object, data, model, usage]
      properties:
        object:
          type: string
          enum: [list]
        data:
          type: array
          items:
            $ref: "#/components/schemas/EmbeddingData"
        model:
          type: string
        usage:
          $ref: "#/components/schemas/EmbeddingUsage"
...
security:
  - ApiKeyHeader: []
  - BearerAuth: []
paths:
...
  /api/generate:
    post:
      operationId: ollamaGenerate
      tags: [Ollama]
      summary: Generate a completion
      description: >
        Generate a completion for the given prompt. When `stream` is true the
        response is NDJSON (application/x-ndjson) with one JSON object per line;
        the final object has `done: true`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OllamaGenerateRequest"
      responses:
        "200":
          description: Completion result (or NDJSON stream when streaming).
          headers:
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OllamaGenerateResponse"
            application/x-ndjson:
              schema:
                $ref: "#/components/schemas/OllamaGenerateResponse"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/chat:
    post:
      operationId: ollamaChat
      tags: [Ollama]
      summary: Chat completion
      description: >
        Send a chat conversation and receive an assistant reply.  Supports
        NDJSON streaming when `stream` is true.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OllamaChatRequest"
      responses:
        "200":
          description: Chat response.
          headers:
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OllamaChatResponse"
            application/x-ndjson:
              schema:
                $ref: "#/components/schemas/OllamaChatResponse"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/pull:
    post:
      operationId: ollamaPull
      tags: [Ollama]
      summary: Pull a model
      description: >
        Download a model from the registry.  Response is NDJSON with progress
        updates.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OllamaPullRequest"
      responses:
        "200":
          description: Pull progress (NDJSON stream).
          headers:
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/x-ndjson:
              schema:
                $ref: "#/components/schemas/OllamaPullProgress"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/show:
    get:
      operationId: ollamaShow
      tags: [Ollama]
      summary: Show model info
      description: Return metadata for a specific model.
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          description: Model name (e.g. "llama3:8b").
      responses:
        "200":
          description: Model information.
          headers:
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShowResponse"
  /api/embeddings:
    post:
      operationId: ollamaEmbeddings
      tags: [Ollama]
      summary: Generate embeddings
      description: Generate an embedding vector for the given prompt.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OllamaEmbeddingsRequest"
      responses:
        "200":
          description: Embedding vector.
          headers:
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OllamaEmbeddingsResponse"

  # ====================================================================
  # OpenAI endpoints
  # ====================================================================
...
  /v1/embeddings:
    post:
      operationId: openaiEmbeddings
      tags: [OpenAI]
      summary: Create embeddings
      description: Generate embedding vectors for the given input.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmbeddingRequest"
      responses:
        "200":
          description: Embedding result.
          headers:
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingResponse"
...